fix bus-off handling (now entirely done in err-warn-int handler)
Not all cases of `bus-off' had been handled, mostly because the err-bus
interrupt was mistakenly interpreted as err-busoff interrupt. So there
was a chance of the SJA being in reset-mode after a bus-off, and
staying there forever, because can4linux didn't realize that condition.
Now bus-off is detected within the err-warn-interrupt handler, and the
reset mode set back if necessary.


imported from: mt@wmipf.in-berlin.de--2005/can4linux--wmipf--0--patch-5